+++ docs_src/scheduler/tasks/from_settings.py
@@ -10,7 +10,7 @@ class AppSettings(EsmeraldAPISettings):
     def scheduler_tasks(self) -> Dict[str, str]:
         return {
             "collect_market_data": "accounts.tasks",
-            "send_email_newsletter": "accounts.tasks",
+            "send_newsletter": "accounts.tasks",
         }