+++ esmerald/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "3.1.5"
+__version__ = "3.2.1"
 
 
 from lilya import status