+++ esmerald/contrib/auth/edgy/base_user.py
@@ -52,7 +52,7 @@ class AbstractUser(edgy.Model):
             self._password = None
             await self.update(password=self.password)
 
-        return check_password(raw_password, str(self.password), setter)
+        return await check_password(raw_password, str(self.password), setter)
 
     async def set_unusable_password(self) -> None:
         # Set a value that will never be a valid hash